+++ packages/cli/devServer.ts
@@ -1,0 +1,5 @@
+import { Server } from '@faasjs/server'
+
+new Server(process.env.FaasRoot as string, {
+  port: Number(process.env.FaasPort as string),
+}).listen()